:truck: Move Member related classes into member package
