+++ src/main/java/com/grepp/coffee/app/controller/web/member/MemberController.java
@@ -4,7 +4,7 @@ package com.grepp.coffee.app.controller.web.member;
 import com.grepp.coffee.app.controller.web.member.payload.SigninRequest;
 import com.grepp.coffee.app.controller.web.member.payload.SignupRequest;
 import com.grepp.coffee.app.model.auth.code.Role;
-import com.grepp.coffee.app.model.dto.MemberDto;
+import com.grepp.coffee.app.model.member.dto.MemberDto;
 import com.grepp.coffee.app.model.member.MemberService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
